Fix broken tests after util->uei rename
